requirements.txt
asgiref==3.7.2
bootstrap4==0.1.0
cart==1.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
Django==4.2
django-appconf==1.0.5
django-crispy-forms==2.0
django-measurement==3.2.4
django-stubs==4.2.3
django-stubs-ext==0.8.0
djangorestframework==3.14.0
idna==3.4
image==1.5.33
measurement==3.2.2
mpmath==1.3.0
mypy==1.2.0
mypy-extensions==1.0.0
Pillow==9.5.0
pycryptodome==3.18.0
Pygments==2.15.1
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
sympy==1.11.1
tomli==2.0.1
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
